Preserve database reservations and saved MPC PSBT inputs during coin selection. Release persisted MPC reservations after broadcast and verify both reservation sources across restart.

+++ tests/contracts/BridgeBase.sol
@@ -34,7 +34,7 @@ abstract contract BridgeBase is Ownable, Pausable {
     event FundsIn(
         address indexed sender,
         uint256 operationId,
-        uint256 amount
+        uint64 amount
     );
 
     // =========================================================================
